refactor type paths
