+++ code/spikes/publish.001/src/common.t.ts
@@ -1,0 +1,2 @@
+import type * as t from './common/types.mjs';
+export type { t };
