+++ code/spikes/publish.001/src/ui/Markdown.Doc/types.mts
@@ -1,4 +1,4 @@
-import type * as t from '../common/types.mjs';
+import type { t } from '../common.t';
 
 type HtmlString = string;
