+++ code/spikes/publish.001/src/ui/Tile.Outline/types.mts
@@ -1,4 +1,4 @@
-import type * as t from '../common/types.mjs';
+import type { t } from '../common.t';
 
 export type RenderTileInner = (args: RenderTileInnerArgs) => JSX.Element | null;
 export type RenderTileInnerArgs = {
