+++ code/spikes/publish.001/src/ui/common/index.ts
@@ -1,4 +1,2 @@
-import type * as t from './types.mjs';
-
-export { t };
+export type { t } from '../common.t';
 export * from './libs.mjs';
